requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
python-dotenv==1.0.1
bcrypt==4.1.2
python-jose==3.3.0
requests==2.31.0 